add file function service, abstractmodel

+++ src/main/java/com/codegym/dao/GenericDAO.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface GenericDAO <T>{
     <T> List<T> query(String sql, RowMapper<T> rowMapper, Object... parameters);
     void update (String sql, Object... parameters);
-    Long insert (String sql, Object... parameters);
+    Integer insert (String sql, Object... parameters);
     int count(String sql, Object... parameters);
 }
